+++ tower-http-client/src/lib.rs
@@ -24,22 +24,26 @@ pub mod request;
 ///
 /// # Examples
 ///
-/// Creating a client and reading the response body using this trait.
-///
-#[doc = include_utils::include_md!("README.md:example")]
+/// ```rust
+#[doc = include_str!("../examples/rate_limiter.rs")]
+/// ```
 ///
 /// [`reqwest::Client`]: https://docs.rs/reqwest/latest/reqwest/struct.Client.html
 pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
     /// Executes an HTTP request.
     fn execute<R>(
-        &self,
+        &mut self,
         request: http::Request<R>,
     ) -> impl Future<Output = Result<http::Response<RespBody>, Err>>
     where
         ReqBody: From<R>;
 
     /// Starts building a request with the given method and URI.
-    fn request<U>(&self, method: Method, uri: U) -> ClientRequest<'_, Self, Err, ReqBody, RespBody>
+    fn request<U>(
+        &mut self,
+        method: Method,
+        uri: U,
+    ) -> ClientRequest<'_, Self, Err, ReqBody, RespBody>
     where
         ReqBody: Default,
         Uri: TryFrom<U>,
@@ -49,7 +53,7 @@ pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
     }
 
     /// Convenience method to make a `GET` request to a given URL.
-    fn get<U>(&self, uri: U) -> ClientRequest<'_, Self, Err, ReqBody, RespBody>
+    fn get<U>(&mut self, uri: U) -> ClientRequest<'_, Self, Err, ReqBody, RespBody>
     where
         ReqBody: Default,
         Uri: TryFrom<U>,
@@ -59,7 +63,7 @@ pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
     }
 
     /// Convenience method to make a `PUT` request to a given URL.
-    fn put<U>(&self, uri: U) -> ClientRequest<'_, Self, Err, ReqBody, RespBody>
+    fn put<U>(&mut self, uri: U) -> ClientRequest<'_, Self, Err, ReqBody, RespBody>
     where
         ReqBody: Default,
         Uri: TryFrom<U>,
@@ -69,7 +73,7 @@ pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
     }
 
     /// Convenience method to make a `POST` request to a given URL.
-    fn post<U>(&self, uri: U) -> ClientRequest<'_, Self, Err, ReqBody, RespBody>
+    fn post<U>(&mut self, uri: U) -> ClientRequest<'_, Self, Err, ReqBody, RespBody>
     where
         ReqBody: Default,
         Uri: TryFrom<U>,
@@ -79,7 +83,7 @@ pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
     }
 
     /// Convenience method to make a `PATCH` request to a given URL.
-    fn patch<U>(&self, uri: U) -> ClientRequest<'_, Self, Err, ReqBody, RespBody>
+    fn patch<U>(&mut self, uri: U) -> ClientRequest<'_, Self, Err, ReqBody, RespBody>
     where
         ReqBody: Default,
         Uri: TryFrom<U>,
@@ -89,7 +93,7 @@ pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
     }
 
     /// Convenience method to make a `DELETE` request to a given URL.
-    fn delete<U>(&self, uri: U) -> ClientRequest<'_, Self, Err, ReqBody, RespBody>
+    fn delete<U>(&mut self, uri: U) -> ClientRequest<'_, Self, Err, ReqBody, RespBody>
     where
         ReqBody: Default,
         Uri: TryFrom<U>,
@@ -99,7 +103,7 @@ pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
     }
 
     /// Convenience method to make a `HEAD` request to a given URL.
-    fn head<U>(&self, uri: U) -> ClientRequest<'_, Self, Err, ReqBody, RespBody>
+    fn head<U>(&mut self, uri: U) -> ClientRequest<'_, Self, Err, ReqBody, RespBody>
     where
         ReqBody: Default,
         Uri: TryFrom<U>,
@@ -111,25 +115,21 @@ pub trait ServiceExt<ReqBody, RespBody, Err>: Sized {
 
 impl<S, ReqBody, RespBody, Err> ServiceExt<ReqBody, RespBody, Err> for S
 where
-    S: Service<http::Request<ReqBody>, Response = http::Response<RespBody>, Error = Err>
-        + Clone
-        + Send
-        + 'static,
+    S: Service<http::Request<ReqBody>, Response = http::Response<RespBody>, Error = Err>,
     S::Future: Send + 'static,
     S::Error: 'static,
 {
-    fn execute<R>(
-        &self,
+    async fn execute<R>(
+        &mut self,
         request: http::Request<R>,
-    ) -> impl Future<Output = Result<http::Response<RespBody>, Err>>
+    ) -> Result<http::Response<RespBody>, Err>
     where
         ReqBody: From<R>,
     {
-        let mut service = self.clone();
-        async move {
-            service.ready().await?;
-            service.call(request.map(ReqBody::from)).await
-        }
+        // Wait until service will be ready to executing requests. It's important for buffered services.
+        self.ready().await?;
+        // And then execute the given request.
+        self.call(request.map(ReqBody::from)).await
     }
 }
 
@@ -146,7 +146,7 @@ pub trait ResponseExt<T>: Sized {
     /// #[tokio::main]
     /// async fn main() -> anyhow::Result<()> {
     ///     // Create a new client
-    ///     let client = HttpClientService::new(reqwest::Client::new());
+    ///     let mut client = HttpClientService::new(reqwest::Client::new());
     ///     // Execute request by using this service.
     ///     let response = client.get("http://ip.jsontest.com").send()?.await?;
     ///